# api-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/general": {
      "get": {
        "operationId": "GeneralController_hello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GetHelloResponseDTO" } } }
          }
        },
        "tags": ["general"]
      }
    },
    "/api/v1/health-check": {
      "get": {
        "operationId": "HealthCheckController_checkHealth",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["health-check"]
      }
    },
    "/api/v1/auth/sign-up": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SignUpBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/sign-in": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [
          {
            "name": "device-name",
            "in": "header",
            "description": "If a mobile device -> send device name and OS",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SignInBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/send-pre-register-sms-code": {
      "post": {
        "operationId": "AuthController_sendPreRegisterSmsCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SendSmsCodeBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/send-forgot-password-sms-code": {
      "post": {
        "operationId": "AuthController_sendForgotPasswordSmsCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SendSmsCodeBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChangePasswordBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/verify-phone": {
      "post": {
        "operationId": "AuthController_verifyPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VerifyPhoneBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/sign-out": {
      "post": {
        "operationId": "AuthController_signOut",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/refresh-tokens": {
      "post": {
        "operationId": "AuthController_refreshTokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/send-feedback": {
      "post": {
        "operationId": "AuthController_sendFeedback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SendFeedbackBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["auth"]
      }
    },
    "/api/v1/aid-kit": {
      "get": {
        "operationId": "AidKitController_getMyAidKits",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["aid-kit"]
      }
    },
    "/api/v1/aid-kit/send-feedback": {
      "post": {
        "operationId": "AidKitController_sendFeedback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SendFeedbackBodyDTO" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["aid-kit"]
      }
    },
    "/api/v1/news/publish": {
      "post": {
        "operationId": "NewsController_sendNews",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateNewsBodyDto" } } }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["news"]
      }
    }
  },
  "info": { "title": "template", "description": "", "version": "0.1.0", "contact": {} },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "GetHelloResponseDTO": {
        "type": "object",
        "properties": { "result": { "type": "string", "example": "string" } },
        "required": ["result"]
      },
      "SignUpBodyDTO": {
        "type": "object",
        "properties": {
          "phone": { "type": "string", "example": "string" },
          "verificationCode": { "type": "number", "example": 1 },
          "password": { "type": "string", "example": "string" },
          "username": { "type": "string", "example": "string" },
          "familyTitle": { "type": "string", "example": "string" },
          "referralCode": { "type": "number", "example": 1 }
        },
        "required": ["phone", "verificationCode", "password", "username", "familyTitle"]
      },
      "SignInBodyDTO": {
        "type": "object",
        "properties": {
          "phone": { "type": "string", "example": "string" },
          "password": { "type": "string", "example": "string" }
        },
        "required": ["phone", "password"]
      },
      "SendSmsCodeBodyDTO": {
        "type": "object",
        "properties": { "phone": { "type": "string", "example": "string" } },
        "required": ["phone"]
      },
      "ChangePasswordBodyDTO": {
        "type": "object",
        "properties": {
          "phone": { "type": "string", "example": "string" },
          "verificationCode": { "type": "number", "example": 1 },
          "newPassword": { "type": "string", "example": "string" },
          "confirmNewPassword": { "type": "string", "example": "string" }
        },
        "required": ["phone", "verificationCode", "newPassword", "confirmNewPassword"]
      },
      "VerifyPhoneBodyDTO": {
        "type": "object",
        "properties": {
          "phone": { "type": "string", "example": "string" },
          "verificationCode": { "type": "number", "example": 1 }
        },
        "required": ["phone", "verificationCode"]
      },
      "RefreshTokenBodyDTO": {
        "type": "object",
        "properties": { "refreshToken": { "type": "string", "example": "string" } }
      },
      "SendFeedbackBodyDTO": {
        "type": "object",
        "properties": { "message": { "type": "string", "example": "string" } },
        "required": ["message"]
      },
      "CreateNewsBodyDto": {
        "type": "object",
        "properties": { "text": { "type": "string", "example": "string" } },
        "required": ["text"]
      }
    }
  }
}